TINY-9633: Create KeyboardNavTabI18n module and begin loading help dialog translations on editor init
