+++ modules/tinymce/src/plugins/help/main/ts/Plugin.ts
@@ -8,6 +8,7 @@ import * as Commands from './api/Commands';
 import * as Options from './api/Options';
 import * as Buttons from './ui/Buttons';
 import * as Dialog from './ui/Dialog';
+import * as KeyboardNavTabI18n from './ui/KeyboardNavTabI18n';
 
 export type TabSpecs = Record<string, DialogType.TabSpec>;
 export type CustomTabSpecs = Cell<TabSpecs>;
@@ -22,6 +23,7 @@ export default (): void => {
     Buttons.register(editor, dialogOpener);
     Commands.register(editor, dialogOpener);
     editor.shortcuts.add('Alt+0', 'Open help dialog', 'mceHelp');
+    KeyboardNavTabI18n.initI18nLoad(editor, pluginUrl);
 
     return api;
   });
